Switch to using double instead of float precision throughout

+++ code/common/audio_file.h
@@ -6,11 +6,11 @@
 
 struct AudioFile {
     size_t NumFrames() const { return interleaved_samples.size() / num_channels; }
-    float &GetSample(unsigned channel, size_t frame) {
+    double &GetSample(unsigned channel, size_t frame) {
         return interleaved_samples[frame * num_channels + channel];
     }
 
-    std::vector<float> interleaved_samples {};
+    std::vector<double> interleaved_samples {};
     unsigned num_channels {};
     unsigned sample_rate {};
     unsigned bits_per_sample = 24;
